modified file structure, BatchID now working
