+++ src/db_user_entry.py
@@ -1,4 +1,4 @@
-#from utils import utils
+from __future__ import print_function
 import sqlite3
 
 username = 'robertej'
